[FE] feat : modal component화
