+++ frontend/moim_front/src/pages/MoimJoinPage/MoimHome/index.jsx
@@ -34,6 +34,7 @@ function MoimHome({ isHost, moimId }) {
 
   const cardClickHandler = (noticeId) => {
     navigate(`notice/${noticeId}`);
+    
   };
 
   const makeNoticeHandler = () => {
